// Capstone Project/src/app/pages/landlord/host-room-detail-page/screen/types.ts
export enum hostRoomDetailGetPageEnum {
  GET_ROOM_DETAIL_REQUEST = 'APP/HOST_HOUSE_LIST/GET_ROOM_DETAIL_REQUEST',
  GET_ROOM_DETAIL_SUCCESS = 'APP/HOST_HOUSE_LIST/GET_ROOM_DETAIL_SUCCESS',
  GET_ROOM_DETAIL_FAILURE = 'APP/HOST_HOUSE_LIST/GET_ROOM_DETAIL_FAILURE',


  UPDATE_ROOM_DETAIL_REQUEST = 'APP/HOST_HOUSE_LIST/UPDATE_ROOM_DETAIL_REQUEST',
  UPDATE_ROOM_DETAIL_SUCCESS = 'APP/HOST_HOUSE_LIST/UPDATE_ROOM_DETAIL_SUCCESS',
  UPDATE_ROOM_DETAIL_FAILURE = 'APP/HOST_HOUSE_LIST/UPDATE_ROOM_DETAIL_FAILURE',


  SET_LOADING_BTN_UPDATE = 'APP/HOST/HOST_HOUSE_LIST/SET_LOADING_BTN_UPDATE',

  CLEAR_STATE = 'APP/HOST/HOST_HOUSE_LIST/CLEAR_STATE',
}